CI/CD: Minor

+++ .ci_cd/deploy.sh
@@ -229,12 +229,15 @@ if [ \$? -ne 0 ]; then
 fi
 
 # Attempt docker compose down
-echo "Stopping existing stack"
-docker-compose -f temp/docker-compose.yml -p or down 2> /dev/null
+CONTAINER_IDS=$(docker ps -q)
+if [ -n "$CONTAINER_IDS" ]; then
+  echo "Stopping existing stack"
+  docker-compose -f temp/docker-compose.yml -p or down 2> /dev/null
 
-if [ \$? -ne 0 ]; then
-  echo "Deployment failed to stop the existing stack"
-  exit 1
+  if [ \$? -ne 0 ]; then
+    echo "Deployment failed to stop the existing stack"
+    exit 1
+  fi
 fi
 
 # Run host init
